actual fix??

+++ project.sql
@@ -10,9 +10,9 @@ TABLE player_all(
     seas_id INT, season NUMERIC(4), player_id NUMERIC, player TEXT, birth_year TEXT,
     pos TEXT, age NUMERIC, experience NUMERIC, lg TEXT, tm TEXT, g NUMERIC, 
     gs NUMERIC, mp NUMERIC, fg NUMERIC, fga NUMERIC,
-    fg_percent FLOAT, x3p NUMERIC, x3pa NUMERIC, x3p_percent NUMERIC(4,1),
-    x2p NUMERIC, x2pa NUMERIC, x2p_percent NUMERIC(4,1), e_fg_percent NUMERIC(4,1),
-    ft NUMERIC, fta NUMERIC, ft_percent NUMERIC(4,1), orb NUMERIC, drb NUMERIC,
+    fg_percent NUMERIC(4,3), x3p NUMERIC, x3pa NUMERIC, x3p_percent NUMERIC(4,3),
+    x2p NUMERIC, x2pa NUMERIC, x2p_percent NUMERIC(4,3), e_fg_percent NUMERIC(4,3),
+    ft NUMERIC, fta NUMERIC, ft_percent NUMERIC(4,3), orb NUMERIC, drb NUMERIC,
     trb NUMERIC, ast NUMERIC, stl NUMERIC, blk NUMERIC, tov NUMERIC, pf NUMERIC, pts NUMERIC
 );
 
@@ -26,9 +26,9 @@ CREATE
 TABLE player_stats(
     seas_id INT, player_id NUMERIC, player TEXT, g NUMERIC, 
     gs NUMERIC, mp NUMERIC, fg NUMERIC, fga NUMERIC,
-    fg_percent NUMERIC, x3p NUMERIC, x3pa NUMERIC, x3p_percent NUMERIC,
-    x2p NUMERIC, x2pa NUMERIC, x2p_percent NUMERIC, e_fg_percent NUMERIC,
-    ft NUMERIC, fta NUMERIC, ft_percent FLOAT, orb NUMERIC, drb NUMERIC,
+    fg_percent NUMERIC(4,3), x3p NUMERIC, x3pa NUMERIC, x3p_percent NUMERIC(4,3),
+    x2p NUMERIC, x2pa NUMERIC, x2p_percent NUMERIC(4,3), e_fg_percent NUMERIC(4,3),
+    ft NUMERIC, fta NUMERIC, ft_percent NUMERIC(4,3), orb NUMERIC, drb NUMERIC,
     trb NUMERIC, ast NUMERIC, stl NUMERIC, blk NUMERIC, tov NUMERIC, pf NUMERIC, pts NUMERIC
 );
